+++ priv/repo/migrations/20230513164357_create_users.exs
@@ -7,7 +7,7 @@ defmodule Inventory.Repo.Migrations.CreateUsers do
       add :name, :string
       add :lastname, :string
       add :age, :integer
-      add :balance, :float
+      add :balance, :decimal
 
       timestamps()
     end
